Move helper classes specific to ListCommand into ListCommand
